Fixed an issue regarding constant session creation
Any request that goes thru the middle ware forces a session creation. this is not ideal as it can spam our memory with session entries. session objects must be created upon handshake only. also It needs to throw whenever the user tries to access an endpoint with swish headers specified. all of this have been addressed by changing the validateRequestSession and loadSession flow and conditions.

Also updated the test tools